Split Handle hierarchy
There are now two parallel Handle hierarchies: global vs connector-specific. SPI operates in terms of the connector-specific hierarchy. The engine (analyzer, planner), operates in terms of the global hierarchy. Metadata/SplitManager/etc take care of wrapping and unwrapping calls and routing to the appropriate connector based on the connectorId embedded in the global handles.

+++ presto-main/src/main/java/com/facebook/presto/execution/DropTableExecution.java
@@ -18,9 +18,11 @@ import com.facebook.presto.metadata.MetadataManager;
 import com.facebook.presto.metadata.NativeTableHandle;
 import com.facebook.presto.metadata.QualifiedTableName;
 import com.facebook.presto.metadata.ShardManager;
+import com.facebook.presto.metadata.TableHandle;
 import com.facebook.presto.metadata.TablePartition;
-import com.facebook.presto.spi.TableHandle;
 import com.facebook.presto.spi.Session;
+import com.facebook.presto.spi.ConnectorTableHandle;
+import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.sql.tree.DropTable;
 import com.facebook.presto.sql.tree.Statement;
 import com.google.common.base.Optional;
@@ -41,7 +43,6 @@ import java.util.concurrent.ThreadPoolExecutor;
 
 import static com.facebook.presto.metadata.MetadataUtil.createQualifiedTableName;
 import static com.facebook.presto.spi.StandardErrorCode.CANNOT_DROP_TABLE;
-import static com.facebook.presto.util.Failures.checkCondition;
 import static com.facebook.presto.util.Threads.daemonThreadsNamed;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
@@ -143,11 +144,14 @@ public class DropTableExecution
 
         final Optional<TableHandle> tableHandle = metadataManager.getTableHandle(tableName);
         checkState(tableHandle.isPresent(), "Table %s does not exist", tableName);
-        checkCondition(tableHandle.get() instanceof NativeTableHandle, CANNOT_DROP_TABLE, "Can only drop native tables");
+        final ConnectorTableHandle connectorHandle = tableHandle.get().getConnectorHandle();
+        if (!(connectorHandle instanceof NativeTableHandle)) {
+            throw new PrestoException(CANNOT_DROP_TABLE.toErrorCode(), "Can only drop native tables");
+        }
 
-        Set<TablePartition> partitions = shardManager.getPartitions(tableHandle.get());
+        Set<TablePartition> partitions = shardManager.getPartitions(tableHandle.get().getConnectorHandle());
         for (TablePartition partition : partitions) {
-            shardManager.dropPartition(tableHandle.get(), partition.getPartitionName());
+            shardManager.dropPartition(connectorHandle, partition.getPartitionName());
         }
 
         metadataManager.dropTable(tableHandle.get());
